requirements.txt
schedule
requests
python-dotenv
duckdb
